+++ code/gopnik.cpp
