+++ code/gopnik.h
