--- main.cpp
@@ -1,11 +1,0 @@
-#include "menu.h"
-#include <QApplication>
-
-int main(int argc, char *argv[])
-{
-    QApplication a(argc, argv);
-    Menu w;
-    w.show();
-
-    return a.exec();
-}
